move 404 to components
